finish splitting
